+++ src/views/Thesaurus.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
 import { computed, onMounted, ref, watch, watchEffect } from "vue";
-import { useStore } from "vuex";
 import { vElementVisibility } from "@vueuse/components";
 import debounce from "lodash/debounce";
-import { key } from "@/store";
+import useRootStore from "@/stores/root.store";
 import useTerms from "@/terms/terms.composable";
 import useTitle from "@/views/title.composable";
 import TermTree from "@/terms/TermTree.vue";
@@ -16,9 +15,9 @@ const PAGE_SIZE = 15;
 const { getRoots, searchTerms } = useTerms();
 const rootTerms = ref<QlitTerm[]>([]);
 const terms = ref<QlitTerm[]>([]);
-const { state } = useStore(key);
+const store = useRootStore();
 useTitle();
-const termTextQuery = computed(() => state.termTextQuery);
+
 const limit = ref(PAGE_SIZE);
 const termsLimited = computed(() => terms.value.slice(0, limit.value));
 const isLoading = ref(false);
@@ -33,21 +32,21 @@ onMounted(async () => {
 // A debounced function for loading terms matching text input
 const findTerms = debounce(async () => {
   isLoading.value = true;
-  terms.value = await searchTerms(termTextQuery.value);
+  terms.value = await searchTerms(store.termTextQuery);
   isLoading.value = false;
 }, 400);
 
 // Reactively determine what terms to show
 watchEffect(async () => {
-  if (termTextQuery.value) {
+  if (store.termTextQuery) {
     // A text query has been entered
-    heading.value = `Ämnen på "${termTextQuery.value}"`;
+    heading.value = `Ämnen på "${store.termTextQuery}"`;
     findTerms();
-  } else if (state.termCollection) {
+  } else if (store.termCollection) {
     // A collection is selected
-    heading.value = state.termCollection.label;
+    heading.value = store.termCollection.label;
     isLoading.value = true;
-    terms.value = await getCollection(state.termCollection.name);
+    terms.value = await getCollection(store.termCollection.name);
     isLoading.value = false;
   } else {
     // Default case: show root terms
@@ -70,7 +69,7 @@ function onBottomVisibility(visible: boolean) {
 <template>
   <div class="container py-6">
     <LoadingSpinner v-if="isLoading" />
-    <div v-else-if="terms.length" :key="termTextQuery">
+    <div v-else-if="terms.length" :key="store.termTextQuery">
       <h1 v-if="heading" class="text-4xl mb-6">{{ heading }}</h1>
       <TermTree
         v-for="term in termsLimited"